Remove duplicate soundcloud embed
In 41d0a1190853d6ae12791c52e692b3afc059664f, a second soundcloud embed
was added. Because all iframes are absolutely positioned (to make the
aspect ratio properly locked, and get a background color behind them),
we probably never saw that the duplicate was introduced.

This removes the one of the two duplicates that has been invisible for
two years, that's probably the safest option by now.